requirement.txt
SQLAlchemy==2.0.36
psycopg2-binary==2.9.10
python-dotenv==1.0.1
sentence-transformers==3.2.1
einops==0.8.0
ollama==0.3.3
fastapi==0.115.4
uvicorn==0.32.0